+++ jaxley/morphology/__init__.py
@@ -1,0 +1,6 @@
+# This file is part of Jaxley, a differentiable neuroscience simulator. Jaxley is
+# licensed under the Apache License Version 2.0, see <https://www.apache.org/licenses/>
+
+from jaxley.morphology.morph_utils import morph_connect, morph_delete
+
+__all__ = ["morph_connect", "morph_delete"]