+++ jaxley/synapses/__init__.py
@@ -5,3 +5,5 @@ from jaxley.synapses.ionotropic import IonotropicSynapse
 from jaxley.synapses.synapse import Synapse
 from jaxley.synapses.tanh_rate import TanhRateSynapse
 from jaxley.synapses.test import TestSynapse
+
+__all__ = ["IonotropicSynapse", "Synapse", "TanhRateSynapse", "TestSynapse"]