+++ jaxley/utils/plot_utils.py
@@ -8,6 +8,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 from matplotlib.axes import Axes
+from matplotlib.patches import Circle
 from numpy import ndarray
 from scipy.spatial import ConvexHull
 
@@ -42,7 +43,18 @@ def plot_graph(
         points = coords_of_branch[:, dims].T
 
         if "line" in type.lower():
-            _ = ax.plot(*points, color=color, **kwargs)
+            std_dev = np.std(coords_of_branch[:, :3], axis=0)
+            if points.shape[1] > 1 and np.any(std_dev > 0.0):
+                _ = ax.plot(*points, color=color, **kwargs)
+            else:
+                # Single point somata are drawn as circles with appropriate radius.
+                circle = Circle(
+                    (points[0, 0], points[1, 0]),
+                    radius=coords_of_branch[0, -1],
+                    facecolor=color,
+                    **kwargs,
+                )
+                _ = ax.add_patch(circle)
         elif "scatter" in type.lower():
             _ = ax.scatter(*points, color=color, **kwargs)
         else: