+++ tests/test_swc.py
@@ -105,16 +105,17 @@ def test_swc_radius(file, swc2jaxley):
 
 
 @pytest.mark.parametrize(
-    "file", ["morph_ca1_n120_single_point_soma.swc", "morph_ca1_n120.swc"]
+    "file",
+    ["morph_ca1_n120_single_point_soma.swc", "morph_ca1_n120.swc"],
 )
-def test_swc_voltages(file, SimpleMorphCell, swc2jaxley):
+def test_swc_voltages(file, SimpleMorphCell):
     """Check if voltages of SWC recording match.
 
     To match the branch indices between NEURON and jaxley, we rely on comparing the
     length of the branches.
 
     It tests whether, on average over time and recordings, the voltage is off by less
-    than 1.5 mV.
+    than 2.0 mV.
     """
     dirname = os.path.dirname(__file__)
     fname = os.path.join(dirname, "swc_files", file)  # n120
@@ -125,7 +126,10 @@ def test_swc_voltages(file, SimpleMorphCell, swc2jaxley):
     t_max = 20.0
     dt = 0.025
 
-    ncomp_per_branch = 8
+    stim_loc = 0.51
+    loc = 0.51
+
+    ncomp_per_branch = 9
 
     ##################### NEURON ##################
     h.secondorder = 0
@@ -144,8 +148,18 @@ def test_swc_voltages(file, SimpleMorphCell, swc2jaxley):
     pathlengths_neuron = np.asarray([sec.L for sec in h.allsec()])
 
     ####################### jaxley ##################
-    _, pathlengths, _, _, _ = swc2jaxley(fname, max_branch_len=2_000)
-    cell = SimpleMorphCell(fname, ncomp_per_branch, max_branch_len=2_000.0)
+    cell = SimpleMorphCell(
+        fname,
+        ncomp_per_branch,
+        max_branch_len=2_000.0,
+        ignore_swc_tracing_interruptions=False,
+    )
+
+    pathlengths = []
+    for branch in cell.branches:
+        pathlengths.append(branch.nodes["length"].sum())
+    pathlengths = np.asarray(pathlengths)
+
     cell.insert(HH())
 
     trunk_inds = [1, 4, 5, 13, 15, 21, 23, 24, 29, 33]
@@ -176,16 +190,16 @@ def test_swc_voltages(file, SimpleMorphCell, swc2jaxley):
     cell.set("HH_h", 0.4889)
     cell.set("HH_n", 0.3644787)
 
-    cell.branch(1).loc(0.05).stimulate(
+    cell.soma.branch(0).loc(stim_loc).stimulate(
         jx.step_current(i_delay, i_dur, i_amp, dt, t_max)
     )
     for i in trunk_inds + tuft_inds + basal_inds:
-        cell.branch(i).loc(0.05).record()
+        cell.branch(i).loc(loc).record()
 
     voltages_jaxley = jx.integrate(cell, delta_t=dt, voltage_solver="jax.sparse")
 
     ################### NEURON #################
-    stim = h.IClamp(h.soma[0](0.1))
+    stim = h.IClamp(h.soma[0](stim_loc))
     stim.delay = i_delay
     stim.dur = i_dur
     stim.amp = i_amp
@@ -197,7 +211,7 @@ def test_swc_voltages(file, SimpleMorphCell, swc2jaxley):
         for i, sec in enumerate(h.allsec()):
             if i == r:
                 v = h.Vector()
-                v.record(sec(0.05)._ref_v)
+                v.record(sec(loc)._ref_v)
                 voltage_recs[f"v{counter}"] = v
                 counter += 1
 
@@ -205,7 +219,7 @@ def test_swc_voltages(file, SimpleMorphCell, swc2jaxley):
         for i, sec in enumerate(h.allsec()):
             if i == r:
                 v = h.Vector()
-                v.record(sec(0.05)._ref_v)
+                v.record(sec(loc)._ref_v)
                 voltage_recs[f"v{counter}"] = v
                 counter += 1
 
@@ -213,7 +227,7 @@ def test_swc_voltages(file, SimpleMorphCell, swc2jaxley):
         for i, sec in enumerate(h.allsec()):
             if i == r:
                 v = h.Vector()
-                v.record(sec(0.05)._ref_v)
+                v.record(sec(loc)._ref_v)
                 voltage_recs[f"v{counter}"] = v
                 counter += 1
 
@@ -246,7 +260,7 @@ def test_swc_voltages(file, SimpleMorphCell, swc2jaxley):
     errors = np.mean(np.abs(voltages_jaxley - voltages_neuron), axis=1)
 
     ####################### check ################
-    assert all(errors < 2.5), "voltages do not match."
+    assert all(errors < 2.0), f"Error {np.max(errors)} > 2.0. Voltages do not match."
 
 
 @pytest.mark.parametrize(